Fixed issues related to displaying ChartView while there is no input data specified

+++ app/src/main/java/com/rd/chartview/view/draw/data/InputData.java
@@ -12,7 +12,7 @@ public class InputData {
 		this.value = value;
 	}
 
-	public InputData(int value, int millis) {
+	public InputData(int value, long millis) {
 		this.value = value;
 		this.millis = millis;
 	}
